* ezEmail : oracle -> mysql(JGw)   - 수신확인/회수, pop3 제외

+++ src/main/java/egovframework/ezEKP/ezEmail/task/EzEmailScheduler.java
@@ -3,8 +3,8 @@ package egovframework.ezEKP.ezEmail.task;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FilenameFilter;
-import java.io.IOException;
 import java.net.URLEncoder;
+import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Date;
 import java.util.List;
@@ -15,7 +15,6 @@ import javax.annotation.Resource;
 import javax.mail.Flags;
 import javax.mail.Folder;
 import javax.mail.Message;
-import javax.mail.MessagingException;
 import javax.mail.Transport;
 import javax.mail.internet.MimeMessage;
 import javax.mail.search.AndTerm;
@@ -25,6 +24,9 @@ import javax.mail.search.ReceivedDateTerm;
 import javax.mail.search.SearchTerm;
 
 import org.apache.commons.lang3.math.NumberUtils;
+import org.json.simple.JSONArray;
+import org.json.simple.JSONObject;
+import org.json.simple.parser.JSONParser;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -73,7 +75,37 @@ public class EzEmailScheduler {
 	public void autoDelete() throws Exception{
 		logger.debug("오전 05:00:00에 호출이 됩니다.");
 		Locale locale = Locale.getDefault();
-		List<MailDeleteVO> list = ezEmailService.getMailDeleteList();
+		
+		List<MailDeleteVO> list = new ArrayList<MailDeleteVO>();
+		
+		if (config.getProperty("config.USE_Mysql").equals("YES")) {
+			String strJson = ezEmailUtil.getWebServiceResult(config.getProperty("config.JGwServerURL") + "/jMochaEzEmail/getMailDeleteAll", null);
+			logger.debug("strJson=" + strJson);
+			
+			JSONParser parser = new JSONParser();
+			JSONObject object = (JSONObject)parser.parse(strJson);
+	        
+	        if (object.get("resultCode").equals("OK") && ((Long)object.get("reasonCode")).intValue() == 0) {
+	        	JSONArray resultArray = (JSONArray)object.get("result");
+	        	
+	        	for (int i=0; i<resultArray.size(); i++) {
+	        		JSONObject obj = (JSONObject)resultArray.get(i);
+	        		
+	        		MailDeleteVO mailDeleteVO = new MailDeleteVO();
+	        		
+	        		mailDeleteVO.setUserId((String)obj.get("userId"));
+	        		mailDeleteVO.setPath((String)obj.get("folderPath"));
+	        		mailDeleteVO.setExpireTime(((Long)obj.get("expireTime")).intValue());
+	        		mailDeleteVO.setDeleteUnread((String)obj.get("deleteUnread"));
+	        		mailDeleteVO.setFolderName((String)obj.get("folderName"));
+	        		
+	        		list.add(mailDeleteVO);
+	        	}
+	        }
+		} else {
+			list = ezEmailService.getMailDeleteList();
+		}
+		
 
 		for (MailDeleteVO vo : list) {
 			IMAPAccess ia = null;
@@ -135,9 +167,34 @@ public class EzEmailScheduler {
 	public void reservedMailSend() throws Exception{
 		logger.debug("reservedMailSend scheduler started.");
 		Locale locale = Locale.getDefault();
-
-		//DB에 있는 예약발송리스트 가져와서 Loop
-		List<MailReservationVO> list = ezEmailService.getMailReserved2();
+		
+		List<MailReservationVO> list = new ArrayList<MailReservationVO>();
+		
+		if (config.getProperty("config.USE_Mysql").equals("YES")) {
+			String strJson = ezEmailUtil.getWebServiceResult(config.getProperty("config.JGwServerURL") + "/jMochaEzEmail/getMailReserved2", null);
+			logger.debug("strJson=" + strJson);
+			
+			JSONParser parser = new JSONParser();
+			JSONObject object = (JSONObject)parser.parse(strJson);
+	        
+	        if (object.get("resultCode").equals("OK") && ((Long)object.get("reasonCode")).intValue() == 0) {
+	        	JSONArray resultArray = (JSONArray)object.get("result");
+	        	
+	        	for (int i=0; i<resultArray.size(); i++) {
+	        		JSONObject obj = (JSONObject)resultArray.get(i);
+	        		
+	        		MailReservationVO vo = new MailReservationVO();
+	        		
+	        		vo.setMessageId((String)obj.get("messageId"));
+	        		vo.setConnUrl((String)obj.get("userId"));
+	        		
+	        		list.add(vo);
+	        	}
+	        }
+		} else {
+			list = ezEmailService.getMailReserved2();
+		}
+		
 		for (MailReservationVO vo : list) {
 			logger.debug(vo.toString());
 			IMAPAccess ia = null;
@@ -186,13 +243,27 @@ public class EzEmailScheduler {
 					f.delete();
 					logger.debug("Succeed in deleting EML file.");
 				} else {
-					logger.debug("Cannot find file. filePath=" + pDirPath + commonUtil.separator + vo.getMessageId() + ".eml");
+					logger.error("Cannot find file. filePath=" + pDirPath + commonUtil.separator + vo.getMessageId() + ".eml");
 				}
-	
-				//DB에서 삭제
-				ezEmailService.deleteMailReserved(vo.getMessageId());
-				logger.debug("Succeed in deleting data from DB.");
 				
+				if (config.getProperty("config.USE_Mysql").equals("YES")) {
+					String inputParams = "messageId=" + URLEncoder.encode(vo.getMessageId(), "UTF-8");
+					logger.debug("inputParams=" + inputParams);
+					
+					String strJson = ezEmailUtil.getWebServiceResult(config.getProperty("config.JGwServerURL") + "/jMochaEzEmail/deleteMailReserved", inputParams);
+					logger.debug("strJson=" + strJson);
+					
+					JSONParser parser = new JSONParser();
+					JSONObject object = (JSONObject)parser.parse(strJson);
+			        
+			        if (!object.get("resultCode").equals("OK") || ((Long)object.get("reasonCode")).intValue() != 0) {
+			        	logger.error("Error from JGwServer");
+			        }
+			        
+				} else {
+					ezEmailService.deleteMailReserved(vo.getMessageId());
+					logger.debug("Succeed in deleting data from DB.");
+				}
 			} catch (Exception e) {
 				e.printStackTrace();
 			} finally {
